PizzaList.java now interacts with the DB.
PizzaList is now working with the DB. The previous errors were due to
the JDBC driver not being installed on the system. They have since been
added to the extension folder under Java which resolved the issue. Also
the topping search had an issue due to it searching for a value called
‘topping’, but in the DB it is called ‘toppings’. This has since been
fixed and is now fully functioning.